refactor metric/retrieval.py to pid_pidgt.py
